sync-github-releases: github-env owns the push-event environment reads
getPushedFiles()/getCommitBeforePush() read the GitHub Actions environment
(GITHUB_EVENT_NAME, GITHUB_SHA, GITHUB_EVENT_PATH), so they belong next to the
other GITHUB_* reads in github-env.ts — not in git.ts. This leaves git.ts as
pure local-git plumbing (every remaining function shells git directly, with no
environment knowledge) and gathers all GitHub Actions environment reads in one
module. gitLines() is now exported for the moved diff call.
